requirements.txt
pandas
numpy
scikit-learn
python-box[all]
ensure
streamlit
pyyaml
python-dotenv
SQLAlchemy

-e .
